\name{Compose}
\alias{Compose}
\title{Compose an arbitrary number of functions.}
\usage{
  Compose(...)
}
\arguments{
  \item{\dots}{the functions to be composed}
}
\value{
  A composed function
}
\description{
  My Happy Hacking keyboard gave out during the writing of
  this procedure; moment of silence, please.
}

