\name{CurryL}
\alias{CurryL}
\title{Lazy curry; thanks, Jamie!
<https://github.com/klutometis/R-functional/issues/1>}
\usage{
  CurryL(FUN, ...)
}
\arguments{
  \item{FUN}{the function to be curried}

  \item{...}{the determining parameters}
}
\description{
  Lazy curry; thanks, Jamie!
  <https://github.com/klutometis/R-functional/issues/1>
}

