% Generated by roxygen2 (4.0.1): do not edit by hand
\name{multi.argument.Compose}
\alias{multi.argument.Compose}
\title{Composition with multiple arguments.}
\usage{
multi.argument.Compose(...)
}
\arguments{
\item{\dots}{the functions to be composed}
}
\value{
A composed function
}
\description{
Thanks, Alexander Davis!
}
\examples{
f <- function(x, y) x+y
g <- function(x) x*2
stopifnot(multi.argument.Compose(f, g)(1,1) == 4)
}

