% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functional.R
\name{Swap}
\alias{Swap}
\title{Thanks, Gabor; see <http://stackoverflow.com/a/23726989>: swaps
the first two arguments in a function.}
\usage{
Swap(f)
}
\arguments{
\item{f}{The function whose arguments to swap}
}
\value{
A function with swapped arguments
}
\description{
Thanks, Gabor; see <http://stackoverflow.com/a/23726989>: swaps
the first two arguments in a function.
}

