% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functional.R
\name{Compose}
\alias{Compose}
\title{Compose an arbitrary number of functions.}
\usage{
Compose(...)
}
\arguments{
\item{\dots}{the functions to be composed}
}
\value{
A composed function
}
\description{
My Happy Hacking keyboard gave out during the writing of this
procedure; moment of silence, please.
}
\details{
Had to make this slightly more complex to handle the
\href{https://github.com/klutometis/R-functional/issues/3}{nullary
case}; also included a fast-path for the trivial case.
}
\examples{
car <- function(list) list[[1]]
cdr <- function(list) list[2:length(list)]
cadr <- Compose(cdr, car)
stopifnot(cadr(c(1,2,3)) == 2)
}

