% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CurryL}
\alias{CurryL}
\title{Lazy curry; thanks, Jamie!
<https://github.com/klutometis/R-functional/issues/1>}
\usage{
CurryL(FUN, ...)
}
\arguments{
\item{FUN}{the function to be curried}

\item{...}{the determining parameters}
}
\description{
Lazy curry; thanks, Jamie!
<https://github.com/klutometis/R-functional/issues/1>
}
\examples{
# day is not defined; thanks, Jamie Folson.
CurryL(function(...) match.call(),
       x=5,
       y=as.Date(day))(z=as.Date(day,"\%Y"))
}

