% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Identity}
\alias{Identity}
\title{Identity function.}
\usage{
Identity(...)
}
\arguments{
\item{\dots}{tautological arguments}
}
\value{
The tautologized arguments, concatenated
}
\description{
Is concatenation benign?
}
\examples{
list.copy <- function(list)
  Reduce(Identity, list)

list <- c(1, 2, 3)
stopifnot(list.copy(list) == list)
}

