\name{Negate}
\alias{Negate}
\title{Negate a function; borrowed from src/library/base/R/funprog.R for
pre-2.7 Rs.}
\usage{
  Negate(f)
}
\arguments{
  \item{f}{the function to be negated}
}
\value{
  The negated function
}
\description{
  Negate a function; borrowed from
  src/library/base/R/funprog.R for pre-2.7 Rs.
}

